changed everything, user sessions are still broken and http sessions are useless

+++ board-server/src/main/java/net/stzups/board/server/websocket/protocol/server/ServerPacketUser.java
@@ -4,7 +4,7 @@ import io.netty.buffer.ByteBuf;
 import net.stzups.board.data.objects.User;
 
 public abstract class ServerPacketUser extends ServerPacket {
-    private int id;
+    private long id;
 
     ServerPacketUser(ServerPacketType packetType, User user) {
         super(packetType);
@@ -14,6 +14,6 @@ public abstract class ServerPacketUser extends ServerPacket {
     @Override
     public void serialize(ByteBuf byteBuf) {
         super.serialize(byteBuf);
-        byteBuf.writeInt(id);
+        byteBuf.writeLong(id);
     }
 }
